[Task-9] added docker files and refactored services and controllers files structure
