+++ Dockerfile
@@ -1,0 +1,13 @@
+FROM node:12.19.0-alpine3.9 AS development
+
+WORKDIR /usr/src/app
+
+COPY package*.json ./
+
+RUN npm install glob rimraf
+
+RUN npm install --only=development
+
+COPY . .
+
+RUN npm run build
