+++ docker-compose.yml
@@ -1,0 +1,23 @@
+version: '3.8'
+
+services:
+  dev:
+    container_name: lesson_service
+    image: nestjs-api-dev:1.0.0
+    build:
+      context: .
+      target: development
+      dockerfile: ./Dockerfile
+    command: npm run start:debug
+    ports:
+      - 3000:3000
+      - 9229:9229
+    networks:
+      - lesson-service
+    volumes:
+      - .:/usr/src/app
+      - /usr/src/app/node_modules
+    restart: unless-stopped
+
+networks:
+  lesson-service:
