+++ package.json
@@ -32,6 +32,7 @@
     "class-validator": "^0.13.2",
     "dotenv": "^16.0.0",
     "install": "^0.13.0",
+    "jest-when": "^3.5.1",
     "mysql2": "^2.3.3",
     "npm": "^8.4.1",
     "reflect-metadata": "^0.1.13",
